+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 * core: Converter will force initialize companion objects to better support `init { Converter.use {...} }`
 * core: Cache keepAlive function introduced to prolong external cache entries
+* server: KeyCipher now uses base64UrlEncode() internally instead of plain base64
 
 # 1.6.5
 * core: simple Cache with expiration timer implemented